Update homepage display
