--- LLMs/requirements.txt
@@ -1,10 +1,0 @@
-flask==2.0.1
-werkzeug==2.0.1
-flask-cors==3.0.10
-gunicorn==20.1.0
-langchain>=0.0.200
-langchain-groq==0.1.10
-openai>=0.27.0
-graphviz>=0.20.1
-opencv-python>=4.5.5.64
-python-dotenv>=0.21.0
